requirements.txt
docker-py==0.4.0
fig==0.5.2
